split off highscore stuff
